[WIP] fix route geometry comparison, extract callbacks into classes, hardcode offline tile pyramid region definition values to match offline region definition provider ones and release a snapshot to test / debug upstream in gl-native
