+++ circle.yml
@@ -130,7 +130,7 @@ jobs:
       - deploy:
           name: Publish Navigation SDK To Maven Central
           command: |
-            if [ "${CIRCLE_BRANCH}" == "master" ]; then
+            if [ "${CIRCLE_BRANCH}" == "dan-offline-maps-ui" ]; then
             make publish ;
             fi
       - store_artifacts:
