+++ libandroid-navigation-ui/src/main/java/com/mapbox/services/android/navigation/ui/v5/NavigationViewModel.java
@@ -331,6 +331,8 @@ public class NavigationViewModel extends AndroidViewModel {
     }
     Context applicationContext = getApplication().getApplicationContext();
     OfflineManager offlineManager = OfflineManager.getInstance(applicationContext);
+    // TODO Tried setting a higher limit
+    //offlineManager.setOfflineMapboxTileCountLimit(1000000);
     //String styleUrl = ThemeSwitcher.retrieveMapStyle(applicationContext);
     // TODO Getting a runtime crash when retrieving the style, hardcoding the styleUrl for testing / debugging purposes
     String styleUrl = "mapbox://styles/mapbox/navigation-guidance-day-v4";
